# Buffer defaults avoiding LOH