--- a/src/NATS.Client.Core/Commands/CommandWriter.cs
+++ b/src/NATS.Client.Core/Commands/CommandWriter.cs
@@ -23,6 +23,9 @@ internal sealed class CommandWriter : IAsyncDisposable
     // assuming 40 bytes TCP overhead + 40 bytes TLS overhead per packet
     private const int SendMemSize = 8520;
 
+    // set to a reasonable socket write mem size
+    private const int MinSegmentSize = 65536;
+
     private readonly ILogger<CommandWriter> _logger;
     private readonly NatsConnection _connection;
     private readonly ObjectPool _pool;
@@ -69,7 +72,7 @@ internal sealed class CommandWriter : IAsyncDisposable
         var pipe = new Pipe(new PipeOptions(
             pauseWriterThreshold: opts.WriterBufferSize, // flush will block after hitting
             resumeWriterThreshold: opts.WriterBufferSize / 2,
-            minimumSegmentSize: opts.WriterBufferSize / 16,
+            minimumSegmentSize: MinSegmentSize,
             useSynchronizationContext: false));
         _pipeReader = pipe.Reader;
         _pipeWriter = pipe.Writer;