--- a/src/NATS.Client.Core/NatsOpts.cs
+++ b/src/NATS.Client.Core/NatsOpts.cs
@@ -31,9 +31,9 @@ public sealed record NatsOpts
 
     public ILoggerFactory LoggerFactory { get; init; } = NullLoggerFactory.Instance;
 
-    public int WriterBufferSize { get; init; } = 1048576;
+    public int WriterBufferSize { get; init; } = 65536;
 
-    public int ReaderBufferSize { get; init; } = 1048576;
+    public int ReaderBufferSize { get; init; } = 65536;
 
     public bool UseThreadPoolCallback { get; init; } = false;