@caleblloyd 2024-02-09
Ran some perf benchmarks, the numbers all came out pretty close.  Doesn't look like anything is way off.

This PR - 64K send/recv buffer size, 16K MinSegmentSize

Test | `main` (msg/s) | This PR (msg/s)
---|---|---
Write Only 50 million msgs, 128B msg size | 3,240,579 | 2,826,737
Write Only 10,000 msgs, 1MiB msg size     |       511 |       479
Read/Write 10 million msgs, 128B msg size | 2,402,587 | 2,535,291
Read/Write 10,000 msgs, 1MiB msg size     |       854 |       882